Welcome to 4.0.0-SNAPSHOT!
Incorporates radical, time-saving and incompatible API updates (but you'll definitely like them), including; removal of *Schema interfaces and classes, removal of Fluent* interfaces and classes, the use of Options everywhere, the introduction of Profiles for applications, the introduction of Extensible types and Features.

Designed to work with Java 7, 8 and 9.  While still a work-in-progress, the foundations are now in place.   It now all builds!